BAN-31 feat: add OAuth signup case when user not registrated via form

+++ apps/users/src/infrastructure/entity/Provider.entity.ts
@@ -7,7 +7,7 @@ import { User } from './User.entity';
 export class Provider extends BaseEntity {
   @Column({ type: 'enum', enum: OauthProviders })
   type: OauthProviders;
-  @Column({ type: 'integer', nullable: true })
+  @Column({ type: 'decimal', nullable: true })
   providerId: number;
   @Column({ type: 'varchar', nullable: true })
   email: string;
